requirements.txt
google-generativeai>=0.3.0
tenacity>=8.2.0
numpy>=1.24.0
pydantic>=2.0.0
python-dateutil>=2.8.2
colorlog>=6.7.0
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
flask>=3.0.0
flask-cors>=4.0.0
